# Oddspark/modules/cms-important-notification/cms-important-notification-impl/rest-openapi.yaml
info:
    description: "CmsImportantNotification REST API"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "CmsImportantNotification"
    version: v1.0
openapi: 3.0.1

components:
    schemas:
        ErrorResponse:
            properties:
                traceId:
                    type: string
                messageInfo:
                    $ref: "#/components/schemas/MessageInfo"
            type: object
        MessageInfo:
            description: "Message information"
            properties:
                messageId:
                    type: string
                messageBody:
                    type: string
            type: object
        DisplayPage:
            description: "Display page information"
            properties:
                key:
                    format: int32
                    type: integer
                name:
                    maxLength: 32
                    type: string
            type: object
        Priority:
            description: "Priority setting information"
            properties:
                id:
                    format: int64
                    type: integer
                displayPages:
                    items:
                        $ref: "#/components/schemas/DisplayPage"
                    type: array
                priority:
                    format: int32
                    type: integer
            type: object



        Item:
            description: "Important notification item"
            properties:
                id:
                    format: int64
                    type: integer
                noticeTitle:
                    maxLength: 128
                    type: string
                noticeContent:
                    maxLength: 255
                    type: string
                importantNoticeStart:
                    description: "Format: YYYY-MM-DD HH:MM:SS"
                    type: string
                importantNoticeEnd:
                    description: "Format: YYYY-MM-DD HH:MM:SS"
                    type: string
                displayBeforeOrAfterLogin:
                    maxLength: 32
                    type: string
                displayPages:
                    items:
                        $ref: "#/components/schemas/DisplayPage"
                    type: array
                priorities:
                    items:
                        $ref: "#/components/schemas/Priority"
                    type: array
            type: object
        ImportantNotificationResponse:
            description: "Response wrapper for important notifications"
            properties:
                items:
                    items:
                        $ref: "#/components/schemas/Item"
                    type: array
                returnedCount:
                    type: integer
                messageInfo:
                    $ref: "#/components/schemas/MessageInfo"
            type: object
paths:
    "/cms-important-notification":
        get:
            description: "Get important notifications by page code"
            operationId: getImportantNotification
            parameters:
                - in: query
                  name: pagecode
                  schema:
                      format: int32
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ImportantNotificationResponse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ImportantNotificationResponse"
                    description: "Success"
                400:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                    description: "Bad Request - Missing pagecode"
                404:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                    description: "Not Found - Invalid pagecode"
                500:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                    description: "Internal Server Error"
                503:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                    description: "Service Unavailable"
            tags: [ "Notification" ]